Make login page match
