+++ app/web/features/landing/locales/en.json
@@ -1,6 +1,6 @@
 {
   "introduction_title": "Meet as strangers, leave as friends.",
-  "introduction_subtitle": "Couchers.org is a community and platform built to facilitate genuine interaction between travelers and locals, centered around couch surfing—free, non-transactional homestays and meetups. Our global community is based on meaningful and inclusive cultural exchange while prioritizing trust and safety.",
+  "introduction_subtitle": "Couchers.org is a community and platform built to facilitate genuine interaction between travelers and locals, centered around <bold>couch surfing</bold>—free, non-transactional homestays and meetups. Our global community is based on meaningful and inclusive cultural exchange while prioritizing trust and safety.",
   "introduction_subtitle2": "Free forever. Community-led. Non-profit. Modern.",
   "signup_header": "Sign up for Couchers.org today",
   "signup_description": "Travel, host, and connect with {{ user_count }} members",
